footer content and style ended, reorganization of certain files(components and img)
