+++ src/pages/Home.js
@@ -1,11 +1,13 @@
-import Navbar from "../components/Navbar/Navbar";
-import Homebody from "../components/Home/Homebody";
+import Navbar from "../components/Navbar";
+import Homebody from "../components/Homebody";
+import Footer from "../components/Footer";
 
 const Home = () => {
   return (
     <div>
       <Navbar />
       <Homebody />
+      <Footer />
     </div>
   );
 };
