+++ src/styles/components/_footer.scss
@@ -1,0 +1,22 @@
+.container {
+  background-color: $black;
+  color: $white;
+  height: 100px;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-items: center;
+  font-size: 10px;
+
+  .sasslogo,
+  .pokeapi {
+    width: 100px;
+    height: 60px;
+  }
+
+  .reactlogo {
+    width: 70px;
+    height: 70px;
+  }
+}
