+++ src/styles/index.scss
@@ -1,3 +1,4 @@
 @import "./settings";
 @import "./components/navbar";
 @import "./components/homebady";
+@import "./components/footer";
